tests(UTItemProviderFileRepresentation): Testing image conversion for kDrive
