+++ Sources/InfomaniakCore/ItemProviderRepresentation/ItemProviderFileRepresentation.swift
@@ -26,12 +26,12 @@ import InfomaniakDI
 public final class ItemProviderFileRepresentation: NSObject, ProgressResultable {
     /// Something to transform events to a nice `async Result`
     private let flowToAsync = FlowToAsyncResult<Success>()
-    
+
     /// Shorthand for default FileManager
     private let fileManager = FileManager.default
-    
+
     /// Domain specific errors
-    public enum ErrorDomain: Error, Equatable{
+    public enum ErrorDomain: Error, Equatable {
         case UTINotFound
         case UnableToLoadFile
     }
@@ -39,7 +39,12 @@ public final class ItemProviderFileRepresentation: NSObject, ProgressResultable 
     public typealias Success = URL
     public typealias Failure = Error
 
-    public init(from itemProvider: NSItemProvider) throws {
+    /// Init method
+    /// - Parameters:
+    ///   - itemProvider: The item provider we will be working with
+    ///   - preferredImageFileFormat: Specify an output image file format. Supports HEIC and JPG. Will convert only if
+    /// itemProvider supports it.
+    public init(from itemProvider: NSItemProvider, preferredImageFileFormat: UTI? = nil) throws {
         guard let typeIdentifier = itemProvider.registeredTypeIdentifiers.first else {
             throw ErrorDomain.UTINotFound
         }
@@ -49,24 +54,31 @@ public final class ItemProviderFileRepresentation: NSObject, ProgressResultable 
 
         super.init()
 
+        // Check if requested an image conversion, and if conversion is available.
+        let fileIdentifierToUse = self.preferredImageFileFormat(
+            itemProvider: itemProvider,
+            typeIdentifier: typeIdentifier,
+            preferredImageFileFormat: preferredImageFileFormat
+        )
+
         // Set progress and hook completion closure to a combine pipe
-        progress = itemProvider.loadFileRepresentation(forTypeIdentifier: typeIdentifier) { [self] fileProviderURL, error in
+        progress = itemProvider.loadFileRepresentation(forTypeIdentifier: fileIdentifierToUse) { [self] fileProviderURL, error in
             guard let fileProviderURL, error == nil else {
                 flowToAsync.sendFailure(error ?? ErrorDomain.UnableToLoadFile)
                 return
             }
 
             do {
-                let UTI = UTI(rawValue: typeIdentifier as CFString)
+                let uti = UTI(rawValue: fileIdentifierToUse as CFString)
                 @InjectService var pathProvider: AppGroupPathProvidable
                 let temporaryURL = pathProvider.tmpDirectoryURL
                     .appendingPathComponent(UUID().uuidString, isDirectory: true)
                 try fileManager.createDirectory(at: temporaryURL, withIntermediateDirectories: true)
 
-                let fileName = fileProviderURL.appendingPathExtension(for: UTI).lastPathComponent
+                let fileName = fileProviderURL.appendingPathExtension(for: uti).lastPathComponent
                 let temporaryFileURL = temporaryURL.appendingPathComponent(fileName)
                 try fileManager.copyItem(atPath: fileProviderURL.path, toPath: temporaryFileURL.path)
-                
+
                 flowToAsync.sendSuccess(temporaryFileURL)
             } catch {
                 flowToAsync.sendFailure(error)
@@ -80,7 +92,35 @@ public final class ItemProviderFileRepresentation: NSObject, ProgressResultable 
 
     public var result: Result<URL, Error> {
         get async {
-            await self.flowToAsync.result
+            await flowToAsync.result
+        }
+    }
+
+    // MARK: Private
+
+    /// Check if a File conversion is possible for the provided `itemProvider` and `typeIdentifier`,
+    /// returns `typeIdentifier` if no conversion is possible.
+    ///
+    /// - Parameters:
+    ///   - itemProvider: The ItemProvider we work with
+    ///   - typeIdentifier: top typeIdentifier for ItemProvider
+    ///   - preferredImageFileFormat: The image format the user is requesting
+    private func preferredImageFileFormat(itemProvider: NSItemProvider,
+                                          typeIdentifier: String,
+                                          preferredImageFileFormat: UTI?) -> String {
+        if let preferredImageFileFormat = preferredImageFileFormat {
+            // Check that itemProvider supports the image types we ask of it
+            if itemProvider.hasItemConformingToAnyOfTypeIdentifiers([UTI.heic.identifier, UTI.jpeg.identifier]),
+               itemProvider.hasItemConformingToTypeIdentifier(preferredImageFileFormat.identifier) {
+                return preferredImageFileFormat.identifier
+            }
+            // No conversion if not possible
+            else {
+                return typeIdentifier
+            }
+        } else {
+            // No conversion
+            return typeIdentifier
         }
     }
 }
